Moved 'Workflows' tab into its own file. Updated the install script using the wrong cuda version. Also updated the dockerfile; untested
