+++ App_Function_Libraries/Gradio_UI/Chat_ui.py
@@ -979,143 +979,6 @@ def create_chat_management_tab():
     return search_query, search_button, conversation_list, conversation_mapping, chat_content, save_button, delete_button, result_message, chat_preview
 
 
-# Load workflows from a JSON file
-json_path = Path('./Helper_Scripts/Workflows/Workflows.json')
-with json_path.open('r') as f:
-    workflows = json.load(f)
-
-
-# FIXME - broken Completely. Doesn't work.
-def chat_workflows_tab():
-    with gr.TabItem("Chat Workflows"):
-        gr.Markdown("# Workflows using LLMs")
-
-        with gr.Row():
-            workflow_selector = gr.Dropdown(label="Select Workflow", choices=[wf['name'] for wf in workflows])
-            api_selector = gr.Dropdown(
-                label="Select API Endpoint",
-                choices=["OpenAI", "Anthropic", "Cohere", "Groq", "DeepSeek", "Mistral", "OpenRouter",
-                         "Llama.cpp", "Kobold", "Ooba", "Tabbyapi", "VLLM", "ollama", "HuggingFace"],
-                value="OpenAI"
-            )
-            api_key_input = gr.Textbox(label="API Key (optional)", type="password")
-
-        context_input = gr.Textbox(label="Initial Context (optional)", lines=5)
-
-        # Create a container for dynamic components
-        with gr.Column() as dynamic_components:
-            prompt_displays = []
-            user_inputs = []
-            output_boxes = []
-            process_buttons = []
-            regenerate_buttons = []
-
-            # Create the maximum number of components needed
-            max_steps = max(len(wf['prompts']) for wf in workflows)
-            for i in range(max_steps):
-                prompt_displays.append(gr.Markdown(visible=False))
-                user_inputs.append(gr.Textbox(label=f"Your Input", lines=2, visible=False))
-                output_boxes.append(gr.Textbox(label=f"AI Output", lines=5, visible=False))
-                with gr.Row():
-                    process_buttons.append(gr.Button(f"Process Step {i + 1}", visible=False))
-                    regenerate_buttons.append(gr.Button(f"🔄 Regenerate", visible=False))
-
-        def update_workflow_ui(workflow_name):
-            selected_workflow = next(wf for wf in workflows if wf['name'] == workflow_name)
-            num_prompts = len(selected_workflow['prompts'])
-
-            prompt_updates = []
-            input_updates = []
-            output_updates = []
-            button_updates = []
-            regenerate_updates = []
-
-            for i in range(max_steps):
-                if i < num_prompts:
-                    prompt_updates.append(
-                        gr.update(value=f"**Step {i + 1}:** {selected_workflow['prompts'][i]}", visible=True))
-                    input_updates.append(gr.update(value="", visible=True, interactive=(i == 0)))
-                    output_updates.append(gr.update(value="", visible=True))
-                    button_updates.append(gr.update(visible=(i == 0)))
-                    regenerate_updates.append(gr.update(visible=False))
-                else:
-                    prompt_updates.append(gr.update(visible=False))
-                    input_updates.append(gr.update(visible=False))
-                    output_updates.append(gr.update(visible=False))
-                    button_updates.append(gr.update(visible=False))
-                    regenerate_updates.append(gr.update(visible=False))
-
-            return prompt_updates + input_updates + output_updates + button_updates + regenerate_updates
-
-        def process(context, user_inputs, workflow_name, api_endpoint, api_key, step):
-            selected_workflow = next(wf for wf in workflows if wf['name'] == workflow_name)
-
-            # Build up the context from previous steps
-            full_context = context + "\n\n"
-            for i in range(step + 1):
-                full_context += f"Question: {selected_workflow['prompts'][i]}\n"
-                full_context += f"Answer: {user_inputs[i]}\n"
-                if i < step:
-                    full_context += f"AI Output: {output_boxes[i].value}\n\n"
-
-            result = process_with_llm(workflow_name, full_context, selected_workflow['prompts'][step], api_endpoint,
-                                      api_key)
-
-            prompt_updates = [gr.update() for _ in range(max_steps)]
-            input_updates = []
-            output_updates = [gr.update() for _ in range(max_steps)]
-            button_updates = []
-            regenerate_updates = []
-
-            for i in range(len(selected_workflow['prompts'])):
-                if i == step:
-                    regenerate_updates.append(gr.update(visible=True))
-                elif i == step + 1:
-                    input_updates.append(gr.update(interactive=True))
-                    button_updates.append(gr.update(visible=True))
-                    regenerate_updates.append(gr.update(visible=False))
-                elif i > step + 1:
-                    input_updates.append(gr.update(interactive=False))
-                    button_updates.append(gr.update(visible=False))
-                    regenerate_updates.append(gr.update(visible=False))
-                else:
-                    input_updates.append(gr.update(interactive=False))
-                    button_updates.append(gr.update(visible=False))
-                    regenerate_updates.append(gr.update(visible=True))
-
-            return [result] + prompt_updates + input_updates + output_updates + button_updates + regenerate_updates
-
-        # Set up event handlers
-        workflow_selector.change(
-            update_workflow_ui,
-            inputs=[workflow_selector],
-            outputs=prompt_displays + user_inputs + output_boxes + process_buttons + regenerate_buttons
-        )
-
-        # Set up process button click events
-        for i, button in enumerate(process_buttons):
-            button.click(
-                fn=lambda context, *user_inputs, wf_name, api_endpoint, api_key, step=i: process(context, user_inputs,
-                                                                                                 wf_name, api_endpoint,
-                                                                                                 api_key, step),
-                inputs=[context_input] + user_inputs + [workflow_selector, api_selector, api_key_input],
-                outputs=[output_boxes[
-                             i]] + prompt_displays + user_inputs + output_boxes + process_buttons + regenerate_buttons
-            )
-
-        # Set up regenerate button click events
-        for i, button in enumerate(regenerate_buttons):
-            button.click(
-                fn=lambda context, *user_inputs, wf_name, api_endpoint, api_key, step=i: process(context, user_inputs,
-                                                                                                 wf_name, api_endpoint,
-                                                                                                 api_key, step),
-                inputs=[context_input] + user_inputs + [workflow_selector, api_selector, api_key_input],
-                outputs=[output_boxes[
-                             i]] + prompt_displays + user_inputs + output_boxes + process_buttons + regenerate_buttons
-            )
-
-    return workflow_selector, api_selector, api_key_input, context_input, dynamic_components
-
 
 # Mock function to simulate LLM processing
 def process_with_llm(workflow, context, prompt, api_endpoint, api_key):
